refactor: useCart 구현 및 관련 연산을 내부로 넣어 개선한다
